docs: Include required file placement in docs
Previously the task execution target file did not have to be in the task root directory; it could be in a subdirectory and be referenced by the `task.json` file. With the new `PowerShell3` task execution handler, this is no longer allowed as it results in a "file not found" error at runtime.