Annotate outStream and errStream as optional
In fact the handling of these properties already allows for them to be undefined/null. But the annotations didn't indicate this.
Since the entire options object was optional, having a couple properties on it *not* optional makes it hard to start supplying *any* options since the user must figure out what values to supply for these two properties to get equivalent behavior, in order to set some *other* option on the object.