Update package.json
Issue #11136 on microsoft/azure-pipelines-tasks points out that sync-request is way more up-to-date than 3.0.1 - Is there any reason not to advance the version for this library? At a minimum changing it to `^3.0.1`, if not something higher?

I personally am experiencing similar issues with this erroring out when being run on a pipeline agent even if my version of package.json has a higher version.