Added unhandledRejection event
The event "unhandledRejection" was added on process.on because when Promise is rejected and there is no catch
the error will not be thrown and process.on('uncaughtException') won't call for Node10 but it will emit "uncaughtException" for node 16.
So we need to listen to "unhandledRejection" and throw the error to make sure the error is thrown and the result will be similar.